update file(修改sort为bigdecimal&& 查询时间段 && 添加引导页 排序)

+++ distribution-parent/distribution-data/src/main/java/com/chauncy/data/domain/po/area/AreaShopLogisticsPo.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -39,7 +40,7 @@ public class AreaShopLogisticsPo implements Serializable {
     private String logiCode;
 
     @ApiModelProperty(value = "排序 越小越靠前")
-    private Integer sort;
+    private BigDecimal sort;
 
 
 }
