+++ distribution-parent/distribution-data/src/main/java/com/chauncy/data/domain/po/store/SmStorePo.java
@@ -66,7 +66,7 @@ public class SmStorePo implements Serializable {
     private Long parentId;
 
     @ApiModelProperty(value = "排序数值")
-    private Integer sort;
+    private BigDecimal sort;
 
     @ApiModelProperty(value = "主理人姓名")
     private String ownerName;
